Added real recipes to DB with ingredients and directions. Modified FoodDetailFragment. READ DESCRIPTION
Looks like a lot of lines but that's because I pasted the script for all 3 script files we have.

Added real recipes with ingredients and directions for our SQL database.
We now have 6 recipes.

Modified the way we print out our food in our FoodDetailActivity.
(a) If image doesn't exist, don't print an image.
(b) Made Ingredients and Directions BOLD and BIGGER

Commented out some if statements in food constructor due to new foods in DB. Whoever wrote them should check it out.

NOTE ONE OF THE TESTS NOW FAIL

testGetFoodBasedOnAnswers fails in our assertTrue part at the end. Probably because we're using the new DB.